refactor: sse 부분 커스텀 훅으로 분리
